+++ R/plotForest.R
@@ -1,14 +1,15 @@
 
+
 #' Forest plot
 #'
 #' Forest plot
 #'
 #' @param x result from \code{dreamlet}
-#' @param coef coefficient to test with \code{topTable}
 #' @param gene gene to show results for
+#' @param coef coefficient to test with \code{topTable}
 #'
 #' @return Plot showing effect sizes
-#'
+#' 
 #' @examples
 #'  
 #' library(muscat)
@@ -34,16 +35,26 @@
 #' coefNames(res.dl)
 #' 
 #' # Show estimated log fold change with in each cell type
-#' plotForest(res.dl, coef = "group_idstim", gene = "ISG20")
+#' plotForest(res.dl, gene = "ISG20", coef = "group_idstim")
 #' 
-#' @import variancePartition limma
-#' @import ggplot2
+#' @rdname plotForest-methods
 #' @export
-plotForest = function(x, coef, gene){
+setGeneric('plotForest', function(x, gene, coef){
+	standardGeneric("plotForest")
+	})
+
+
+
+
+
+#' @rdname plotForest-methods
+#' @aliases plotForest,dreamletResult-method
+#' @export
+setMethod("plotForest", signature(x="dreamletResult"),
+	function(x, gene, coef){
 
 	# Pass R CMD check
 	Assay = logFC = FDR = se = NULL
-
 
 	df = lapply( names(x), function(assay){
 		tab = topTable(x[[assay]], coef=coef, number=Inf)
@@ -52,9 +63,46 @@ plotForest = function(x, coef, gene){
 	df = do.call(rbind, df)
 	df$FDR = p.adjust(df$P.Value)
 
-	ggplot(df[df$Gene == gene, ], aes(Assay, logFC,  color=-log10(pmax(1e-4,FDR)) )) + geom_point() + geom_errorbar(aes(ymin = logFC - 1.96*se, ymax = logFC + 1.96*se), width=0.1) + theme_classic() + theme(plot.title = element_text(hjust = 0.5)) + coord_flip() + ggtitle(gene) + ylab(bquote(log[2]~fold~change)) + geom_hline(yintercept=0, linetype="dashed") + scale_color_gradient(name = bquote(-log[10]~FDR), low="grey", high="red", limits=c(0, 4))
-}
+	ggplot(df[df$Gene == gene, ], aes(Assay, logFC,  color=-log10(pmax(1e-4,FDR)) )) + geom_point() + geom_errorbar(aes(ymin = logFC - 1.96*se, ymax = logFC + 1.96*se), width=0.1) + theme_classic() + theme(plot.title = element_text(hjust = 0.5)) + coord_flip() + ggtitle(gene) + ylab(bquote(log[2]~fold~change)) + geom_hline(yintercept=0, linetype="dashed") + scale_color_gradient(name = bquote(-log[10]~FDR), low="grey", high="red", limits=c(0, 4)) + xlab('')
+})
 
+
+
+
+
+#' @rdname plotForest-methods
+#' @aliases plotForest,dreamlet_mash_result-method
+#' @importFrom ashr get_pm get_lfsr get_psd
+#' @export
+setMethod("plotForest", signature(x="dreamlet_mash_result"),
+	function(x, gene, coef){
+
+	# Pass R CMD check
+	ID = logFC = lFSR = se = NULL
+
+	# extract mashr statistics
+	df_logFC = reshape2::melt(get_pm(x$model))
+	colnames(df_logFC) = c("Gene", "ID", "logFC")
+	df_logFC$key = paste(df_logFC$Gene, df_logFC$ID)
+
+	df_lfsr = reshape2::melt(get_lfsr(x$model))
+	colnames(df_lfsr) = c("Gene", "ID", "lFSR")
+	df_lfsr$key = paste(df_lfsr$Gene, df_lfsr$ID)
+
+	df_se = reshape2::melt(get_psd(x$model))
+	colnames(df_se) = c("Gene", "ID", "se")
+	df_se$key = paste(df_se$Gene, df_se$ID)
+
+	# merge mashr statistics
+	df = merge(df_logFC, df_lfsr, by='key')
+	df = merge(df, df_se, by='key')
+
+	# drop NA values
+	df = df[!is.na(df$logFC),]
+
+	# make plot
+	ggplot(df[df$Gene.x == gene, ], aes(ID, logFC,  color=-log10(pmax(1e-4,lFSR)) )) + geom_point() + geom_errorbar(aes(ymin = logFC - 1.96*se, ymax = logFC + 1.96*se), width=0.1) + theme_classic() + theme(plot.title = element_text(hjust = 0.5)) + coord_flip() + ggtitle(gene) + ylab(bquote(log[2]~fold~change)) + geom_hline(yintercept=0, linetype="dashed") + scale_color_gradient(name = bquote(-log[10]~lFSR), low="grey", high="red", limits=c(0, 4)) + xlab('')
+})